moved user to blueprints
